Ensure BottomSheets call ButterKnife unbinder
